Add `@moduletag :error_view_case` to `ErrorViewTest`
### Running Tests Using Tags

ExUnit allows us to tag our tests at the case level or on the individual test level. We can then choose to run only the tests with a specific tag, or we can exclude tests with that tag and run everything else.

Let's experiment with how this works.

First, we'll add a `@moduletag` to `test/views/error_view_test.exs`.

    # test/views/error_view_test.exs
    defmodule HelloPhoenix.ErrorViewTest do
      use HelloPhoenix.ConnCase, async: true

      @moduletag :error_view_case
      ...
    end

If we use only an atom for our module tag, ExUnit assumes that it has a value of `true`. We could also specify a different value if we wanted.

    # test/views/error_view_test.exs
    defmodule HelloPhoenix.ErrorViewTest do
      use HelloPhoenix.ConnCase, async: true

      @moduletag error_view_case: "some_interesting_value"
      ...
    end

For now, let's leave it as a simple atom `@moduletag :error_view_case`.

We can run only the tests from the error view case by passing `--only error_view_case` into `mix test`.

    $ mix test --only error_view_case
    Including tags: [:error_view_case]
    Excluding tags: [:test]

    \...

    Finished in 0.1 seconds (0.1s on load, 0.00s on tests)
    3 tests, 0 failures

    Randomized with seed 125659

> Note: ExUnit tells us exactly which tags it is including and excluding for each test run. If we look back to the previous section on running tests, we'll see that line numbers specified for individual tests are actually treated as tags.

    $ mix test test/views/error_view_test.exs:12
    Including tags: [line: "12"]
    Excluding tags: [:test]

    \.

    Finished in 0.2 seconds (0.2s on load, 0.00s on tests)
    1 tests, 0 failures

    Randomized with seed 364723

Specifying a value of `true` for `error_view_case` yields the same results.

    $ mix test --only error_view_case:true
    Including tags: [error_view_case: "true"]
    Excluding tags: [:test]

    \...

    Finished in 0.1 seconds (0.1s on load, 0.00s on tests)
    3 tests, 0 failures

    Randomized with seed 833356

Specifying `false` as the value for `error_view_case`, however, will not run any tests because no tags in our system match `error_view_case: false`.

    $ mix test --only error_view_case:false
    Including tags: [error_view_case: "false"]
    Excluding tags: [:test]

    \

    Finished in 0.1 seconds (0.1s on load, 0.00s on tests)
    0 tests, 0 failures

    Randomized with seed 622422

We can use the `--exclude` flag in a similar way. This will run all of the tests except those in the error view case.

    $ mix test --exclude error_view_case
    Excluding tags: [:error_view_case]

    \.

    Finished in 0.2 seconds (0.1s on load, 0.03s on tests)
    1 tests, 0 failures

    Randomized with seed 682868

Specifying values for a tag works the same way for `--exclude` as it does for `--only`.