Add two tags to `HelloPhoenix.ErrorViewTest`
We can tag individual tests as well as full test cases. Let's tag a few
tests in the error view case to see how this works.

    # test/views/error_view_test.exs
    defmodule HelloPhoenix.ErrorViewTest do
      use HelloPhoenix.ConnCase, async: true

      @moduletag :error_view_case

      # Bring render/3 and render_to_string/3 for testing custom views
      import Phoenix.View

      @tag individual_test: "yup"
      test "renders 404.html" do
        assert render_to_string(HelloPhoenix.ErrorView, "404.html", []) ==
               "Page not found"
      end

      @tag individual_test: "nope"
      test "render 500.html" do
        assert render_to_string(HelloPhoenix.ErrorView, "500.html", []) ==
               "Server internal error"
      end
      ...
    end

If we would like to run only tests tagged as `individual_test`,
regardless of their value, this will work.

    $ mix test --only individual_test
    Including tags: [:individual_test]
    Excluding tags: [:test]

    ..

    Finished in 0.1 seconds (0.1s on load, 0.00s on tests)
    2 tests, 0 failures

    Randomized with seed 813729

We can also specify a value and run only tests with that.

    $ mix test --only individual_test:yup
    Including tags: [individual_test: "yup"]
    Excluding tags: [:test]

    .

    Finished in 0.1 seconds (0.1s on load, 0.00s on tests)
    1 tests, 0 failures

    Randomized with seed 770938

Similarly, we can run all tests except for those tagged with a given value.

    $ mix test --exclude individual_test:nope
    Excluding tags: [individual_test: "nope"]

    ...

    Finished in 0.2 seconds (0.1s on load, 0.03s on tests)
    3 tests, 0 failures

    Randomized with seed 539324

Here's an interesting case, let's exclude all of the tests from the error view case except the ones tagged with `individual_test`.

    $ mix test --exclude error_view_case --include individual_test
    Including tags: [:individual_test]
    Excluding tags: [:error_view_case]

    ...

    Finished in 0.2 seconds (0.1s on load, 0.03s on tests)
    3 tests, 0 failures

    Randomized with seed 41241

This runs the two tests tagged with `individual_test` as well as the one from `test/controllers/page_controller_test.exs`.

We can be more specific and exclude all the tests from the error view case except the one tagged with `individual_test` that has the value "yup".

    $ mix test --exclude error_view_case --include individual_test:yup
    Including tags: [individual_test: "yup"]
    Excluding tags: [:error_view_case]

    ..

    Finished in 0.2 seconds (0.1s on load, 0.03s on tests)
    2 tests, 0 failures

    Randomized with seed 61472