Revert "Exclude `error_view_case: true` from `test_helper.exs`"
### Randomization

Running tests in random order is a good way to ensure that our tests are truly isolated. If we notice that we get sporadic failures for a given test, it may be because a previous test changes the state of the system in ways that aren't cleaned up afterward, thereby affecting the tests which follow. Those failures might only present themselves if the tests are run in a specific order.

ExUnit, will randomize the order tests run in by default, using an integer to seed the randomization. If we notice that a specific random seed triggers our intermittent failure, we can re-run the tests with that same seed to reliably recreate that test sequence in order to help us figure out what the problem is.

    $ mix test --seed 401472
    \....

    Finished in 0.2 seconds (0.1s on load, 0.03s on tests)
    4 tests, 0 failures

    Randomized with seed 401472

### Generating More Files

We've seen what Phoenix gives us with a newly generated app. Now let's
see what happens when we generate a new HTML resource.

Let's borrow the `users` resource we created in the
[Ecto Models Guide](http://www.phoenixframework.org/docs/ecto-models).

At the root of our new application, let's run the `mix phoenix.gen.html`
task with the following options.

    $ mix phoenix.gen.html User users name:string email:string bio:string number_of_pets:integer
    * creating web/controllers/user_controller.ex
    * creating web/templates/user/edit.html.eex
    * creating web/templates/user/form.html.eex
    * creating web/templates/user/index.html.eex
    * creating web/templates/user/new.html.eex
    * creating web/templates/user/show.html.eex
    * creating web/views/user_view.ex
    * creating test/controllers/user_controller_test.exs
    * creating priv/repo/migrations/20151012202643_create_user.exs
    * creating web/models/user.ex
    * creating test/models/user_test.exs

    Add the resource to your browser scope in web/router.ex:

        resources "/users", UserController

    Remember to update your repository by running migrations:

        $ mix ecto.migrate