Add the "/users" resources to `web/router.ex`
Now let's follow the directions and add the new resources route to our
`web/router.ex` file.

    # web/router.ex
    defmodule HelloPhoenix.Router do
      ...

      scope "/", HelloPhoenix do
        pipe_through :browser # Use the default browser stack

        get "/", PageController, :index
        resources "/users", UserController
      end

      # Other scopes may use custom stacks.
      # scope "/api", HelloPhoenix do
      #   pipe_through :api
      # end
    end

When we run `mix test` again, we see that we already have sixteen tests!

    $ mix test
    Compiled lib/hello_phoenix.ex
    Compiled web/models/user.ex
    Compiled web/views/page_view.ex
    Compiled web/views/layout_view.ex
    Compiled web/views/error_view.ex
    Compiled web/controllers/page_controller.ex
    Compiled web/router.ex
    Compiled web/controllers/user_controller.ex
    Compiled web/views/user_view.ex
    Compiled lib/hello_phoenix/endpoint.ex
    Generated hello_phoenix app
    ................

    Finished in 0.5 seconds (0.4s on load, 0.1s on tests)
    16 tests, 0 failures

    Randomized with seed 537537

At this point, we are at a great place to transition to the rest of the
testing guides, in which we'll examine these tests in much more detail,
and add some of our own.

# Models

In the [Ecto Models Guide](http://www.phoenixframework.org/docs/ecto-models)
we generated an HTML resource for users. This gave us a number of
modules for free, including a user model and a user model test case. In
this guide, we'll use the model and test case to work through the
changes we made in the Ecto Models Guide in a test-driven way.

For those of us who haven't worked through the Ecto Models Guide, it's
easy to catch up. Please see the "Generating an HTML Resource" section
below.

Before we do anything else, let's run `mix test` to make sure our test
suite runs cleanly.

    $ mix test
    ................

    Finished in 0.6 seconds (0.5s on load, 0.1s on tests)
    16 tests, 0 failures

    Randomized with seed 638414

Great. We've got sixteen tests and they are all passing!

## Test Driving a Changeset

The focus of this guide is going to be on `test/models/user_test.exs`.
Let's take a quick look to get familiar with it.

    # test/models/user_test.exs
    defmodule HelloPhoenix.UserTest do
      use HelloPhoenix.ModelCase

      alias HelloPhoenix.User

      @valid_attrs %{bio: "some content", email: "some content", name: "some content", number_of_pets: 42}
      @invalid_attrs %{}

      test "changeset with valid attributes" do
        changeset = User.changeset(%User{}, @valid_attrs)
        assert changeset.valid?
      end

      test "changeset with invalid attributes" do
        changeset = User.changeset(%User{}, @invalid_attrs)
        refute changeset.valid?
      end
    end

In the first line, we `use HelloPhoenix.ModelCase`, which is defined in
`test/support/model_case.ex`. `HelloPhoenix.ModelCase` is responsible
for importing and aliasing all the necessary modules for all of our
model cases. `HelloPhoenix.ModelCase` will also run all of our model
tests within a database transaction unless we've tagged an individual
test case with `:async`.

> Note: We should not tag any model case that interacts with a database
as `:async`. This may cause  erratic test results and possibly even deadlocks.

`HelloPhoenix.ModelCase` is also a place to define any helper functions
we might need to test our models. We get an example function
`errors_on/2` for free, and we'll see how that works shortly.

We alias our `HelloPhoenix.User` module so that we can refer to its
structs as `%User{}` instead of `%HelloPhoenix.User{}`.

We also define module attributes for `@valid_attrs` and `@invalid_attrs`
so they will be available to all our tests.