Update `@valid_attrs` in `HelloPhoenix.UserTest`
The generated test attributes we get from `HelloPhoenix.UserTest` are
certainly usable as is, but let's change them to look just a bit more
realistic. The only one that will really matter is `:email`, as that
will need to have an `@` before we're done. The other changes are just
cosmetic.

    # test/models/user_test.exs
    defmodule HelloPhoenix.UserTest do
      use HelloPhoenix.ModelCase

      alias HelloPhoenix.User

      @valid_attrs %{bio: "my life", email: "pat@example.com", name: "Pat Example", number_of_pets: 4}
      @invalid_attrs %{}

      ...
    end

We should change the `@valid_attrs` module attribute in
`test/controllers/user_controller_test.exs` to match these as well for
consistency.

    # test/controllers/user_controller_test.exs
    defmodule HelloPhoenix.UserControllerTest do
      use HelloPhoenix.ConnCase

      alias HelloPhoenix.User
      @valid_attrs %{bio: "my life", email: "pat@example.com", name: "Pat Example", number_of_pets: 4}
      @invalid_attrs %{}

      ...
    end

If we run the tests again, all sixteen should still pass.