Add a failing test in `HelloPhoenix.UserTest`
#### Number of Pets

While Phoenix generated our model with all of the fields required, the
number of pets a user has is optional in our domain.

Let's write a new test to verify that.

To test this, we can delete the `:number_of_pets` key and value from the
`@valid_attrs` map and make a `User` changeset from those new
attributes. Then we can assert that the changeset is still valid.

    # test/models/user_test.exs
    defmodule HelloPhoenix.UserTest do
      ...

      test "number_of_pets is not required" do
        changeset = User.changeset(%User{}, Map.delete(@valid_attrs, :number_of_pets))
        assert changeset.valid?
      end
    end

Now, let's run the tests again.

    $ mix test
    .............

    1) test number_of_pets is not required (HelloPhoenix.UserTest)
       test/models/user_test.exs:19
       Expected truthy, got false
       code: changeset.valid?()
       stacktrace:
         test/models/user_test.exs:21

    ..

    Finished in 0.4 seconds (0.2s on load, 0.1s on tests)
    17 tests, 1 failures

    Randomized with seed 780208