Add a failing test for `User.bio`
#### The Bio Attribute

In the Ecto Models Guide, we learned that the user's `:bio` attribute
has two business requirements. The first is that it must be at least
two characters long. Let's write a test for that using the same pattern
we've just used.

First, we change the `:bio` attribute to have a value of a single
character. Then we create a changeset with the new attributes and test
its validity.

    # test/models/user_test.exs
    defmodule HelloPhoenix.UserTest do
      ...

      test "bio must be at least two characters long" do
        attrs = %{@valid_attrs | bio: "I"}
        changeset = User.changeset(%User{}, attrs)
        refute changeset.valid?
      end
    end

When we run the test, it fails, as we would expect.

    $ mix test
    .....

      1) test bio must be at least two characters long (HelloPhoenix.UserTest)
         test/models/user_test.exs:24
         Expected false or nil, got true
         code: changeset.valid?()
         stacktrace:
           test/models/user_test.exs:27

    ............

    Finished in 0.4 seconds (0.3s on load, 0.1s on tests)
    18 tests, 1 failures

    Randomized with seed 231093

Hmmm. Yes, this test behaved as we expected, but the error message
doesn't seems reflect our test. We're validating the length of the
`:bio` attribute, and the message we get is "Expected false or nil, got
true". There's no mention of our `:bio` attribute at all.