Assert an error in `UserTest` instead of refuting `valid?`
We can do better.

Let's change our test to get a better message while still testing the same behavior. We can leave the code to set the new `:bio` value in place. In the `assert`, however, we'll use the `errors_on/2` function we get from `ModelCase` to generate a list of errors, and check that the `:bio` attribute error is in that list.

    # test/models/user_test.exs
    defmodule HelloPhoenix.UserTest do
      ...

      test "bio must be at least two characters long" do
        attrs = %{@valid_attrs | bio: "I"}
        assert {:bio, {"should be at least %{count} characters", [count: 2]}} in errors_on(%User{}, attrs)
      end
    end

> Note: `ModuleCase.errors_on/2` returns a keyword list, and an individual element of a keyword list is a tuple.

When we run the tests again, we get a different message entirely.

    $ mix test
    ....

      1) test bio must be at least two characters long (HelloPhoenix.UserTest)
         test/models/user_test.exs:24
         Assertion with in failed
         code: {:bio, {"should be at least %{count} characters", [count: 2]}} in errors_on(%User{}, attrs)
         lhs:  {:bio, {"should be at least %{count} characters", [count: 2]}}
         rhs:  []
         stacktrace:
           test/models/user_test.exs:26

    .............

    Finished in 0.7 seconds (0.4s on load, 0.2s on tests)
    18 tests, 1 failures

    Randomized with seed 740790

This shows us the assertion we are testing - that our error is in the list of errors from the model's changeset.

    code: {"should be at least %{count} characters", [count: 2]} in errors_on(%User{}, attrs)

We see that the left hand side of the expression evaluates to our error.

    lhs:  {"should be at least %{count} characters", [count: 2]}

And we see that the right hand side of the expression evaluates to an empty list.

    rhs:  []

That list is empty because we don't yet validate the minimum length of the `:bio` attribute.