Add another failing test for `User.bio`
The other business requirement for the `:bio` field is that it be a
maximum of one hundred and forty characters. Let's write a test for that
using the `errors_on/2` function again.

Before we actually write the test, how are we going to handle a string
that long without making a mess? A new function in
`HelloPhoenix.ModelCase` is perfect for this. We'll create a
`long_string/1` function which will send us back a string of "a"'s as
long as we tell it to be.

    # test/support/model_case.ex
    defmodule HelloPhoenix.ModelCase do
      ...

      def long_string(length) do
        Enum.reduce (1..length), "", fn _, acc ->  acc <> "a" end
      end
    end

We can now use `long_string/1` when changing the value of the `:bio` key
in our `attrs`.

    # test/models/user_test.exs
    defmodule HelloPhoenix.UserTest do
      ...

      test "bio must be at most 140 characters long" do
        attrs = %{@valid_attrs | bio: long_string(141)}
        assert {:bio, {"should be at most %{count} characters", [count: 140]}} in errors_on(%User{}, attrs)
      end
    end

When we run the test, it fails as we want it to.

    $ touch test/support/model_case.ex; mix test
    ....

      1) test bio must be at most 140 characters long (HelloPhoenix.UserTest)
         test/models/user_test.exs:29
         Assertion with in failed
         code: {:bio, {"should be at most %{count} characters", [count: 140]}} in errors_on(%User{}, attrs)
         lhs:  {:bio, {"should be at most %{count} characters", [count: 140]}}
         rhs:  []
         stacktrace:
           test/models/user_test.exs:31

    ..............

    Finished in 0.5 seconds (0.3s on load, 0.1s on tests)
    19 tests, 1 failures

    Randomized with seed 331812