Add a failing test for `User.email`
#### The Email Attribute

We have one last attribute to validate. Currently, `:email` is just a string like any other. We'd like to make sure that it at least matches an "@". This is no substitute for an email confirmation, but it will weed out some invalid addresses before we even try.

This process will feel familiar by now. First, we change the value of the `:email` attribute to omit the "@". Then we write an assertion which uses `errors_on/2` to check for the correct validation error on the `:email` attribute.

    # test/models/user_test.exs
    defmodule HelloPhoenix.UserTest do
      ...

      test "email must contain at least an @" do
        attrs = %{@valid_attrs | email: "fooexample.com"}
        assert {:email, "has invalid format"} in errors_on(%User{}, attrs)
      end
    end

When we run the tests, it fails. We see that we're getting an empty list of errors back from `errors_on/2`.

    $ mix test
    \................

    1) test email must contain at least an @ sign (HelloPhoenix.UserTest)
    test/models/user_test.exs:34
    Assertion with in failed
    code: {:email, "has invalid format"} in errors_on(%User{}, attrs)
      lhs:  {:email, "has invalid format"}
      rhs:  []
      stacktrace:
      test/models/user_test.exs:38

      \..

    Finished in 0.4 seconds (0.2s on load, 0.1s on tests)
    19 tests, 1 failures

    Randomized with seed 962127